feat: first commit
A small demo of how to use the openfeature Go library in a containerised
(kubernetes) application. For the purposes of the demo it loops and
tries to check a feature flag every 2s ("bake a cake").

We have to wait for the openfeature provider to initialise, there are no
helper functions for this in the openfeature library that I can see, so
I have achieved this by listening to the ProviderReady event and
synchronising using a channel which seems to work well.